requirements.txt
Django==5.0.3
django-cors-headers==4.3.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
packaging==24.0
pycryptodome==3.20.0
asgiref==3.8.1
PyJWT==2.8.0
sqlparse==0.4.4
typing_extensions==4.10.0
gunicorn==21.2.0
